refactor to make upcoming testing easier
